flask/requirements.txt
attrs==21.4.0
blinker==1.4
coverage==6.3.1
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-pymysql==0.2.3
google-auth==2.3.3
image==1.5.33
iniconfig==1.1.1
itsdangerous==2.0.1
packaging==21.3
pika==1.2.0
pluggy==1.0.0
pusher==3.2.0
PyJWT==2.3.0
pyotp==2.6.0
pyparsing==3.0.7
pytest==7.0.1
pytest-mock==3.7.0
python-dotenv==0.19.2
pytz==2021.3
qrcode==7.3.1
requests==2.27.1
rollbar==0.16.2
tomli==2.0.1
Werkzeug==2.0.2
waitress==2.1.1
